feat: use "md5sum -b" style separator
This commit changes the format of the generated checksum file
to be as if generated by `md5sum -b`. The `-b` flag reads the input
in binary mode. The checksum is read in binary mode so this change
reflects what is actually the case.

In addition this will simplify the WARC-file pipeline downstream
because our storage system expects checksums generated with binary
mode enabled. Right now we have to change the separator after the
fact.